Refactor time scale methods into a common location
